split up tests into multiple files
